Remove wallet SignAny from deprecated
